add new prompt done, restructure folders
