+++ src/mish/init_mish.c
